Implement status update methods and event listener for checkbox
