+++ src/modules/localStorage.js
@@ -1,0 +1,10 @@
+function getTasks() {
+  const tasks = localStorage.getItem('tasks');
+  return tasks ? JSON.parse(tasks) : [];
+}
+
+function saveTasks(tasks) {
+  localStorage.setItem('tasks', JSON.stringify(tasks));
+}
+
+export { getTasks, saveTasks };
